Commit changing tests

+++ src/main/java/com/vironit/kazimirov/service/impl/PurposeServiceImpl.java
@@ -1,5 +1,8 @@
 package com.vironit.kazimirov.service.impl;
 
+import com.vironit.kazimirov.entity.Good;
+import com.vironit.kazimirov.exception.CantDeleteElement;
+import com.vironit.kazimirov.fakedao.DaoInterface.GoodDao;
 import com.vironit.kazimirov.fakedao.DaoInterface.PurposeDao;
 import com.vironit.kazimirov.entity.Purpose;
 import com.vironit.kazimirov.exception.RepeatitionException;
@@ -8,20 +11,30 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PurposeServiceImpl implements PurposeService {
     @Autowired
     private final PurposeDao purposeDao;
+    @Autowired
+    private final GoodDao goodDao;
 
     @Autowired
-    public PurposeServiceImpl(PurposeDao purposeDao) {
+    public PurposeServiceImpl(PurposeDao purposeDao, GoodDao goodDao) {
         this.purposeDao = purposeDao;
+        this.goodDao = goodDao;
     }
 
     @Override
     public void addPurpose(Purpose purpose) throws RepeatitionException {
-        purposeDao.addPurpose(purpose);
+        Optional<Purpose> checkPurpose = Optional.ofNullable(purposeDao.findPurposeByName(purpose.getPurpose()));
+        if (checkPurpose.isPresent()==false) {
+            purposeDao.addPurpose(purpose);
+        } else {
+            throw new RepeatitionException("such purpose is present");
+        }
+        //purposeDao.addPurpose(purpose);
     }
 
     @Override
@@ -35,8 +48,21 @@ public class PurposeServiceImpl implements PurposeService {
     }
 
     @Override
-    public void deletePurpose(int idPurpose) {
-        purposeDao.deletePurpose(idPurpose);
+    public void deletePurpose(int idPurpose) throws CantDeleteElement {
+        Purpose purpose=purposeDao.findPurposeById(idPurpose);
+        List<Good> goods=goodDao.findByPurpose(purpose);
+        if (goods.size()==0) {
+            purposeDao.deletePurpose(idPurpose);
+        } else {
+            throw new CantDeleteElement("this purpose is using with goods");
+        }
+        //purposeDao.deletePurpose(idPurpose);
     }
+
+    @Override
+    public Purpose findPurposeById(int idPurpose) {
+        return purposeDao.findPurposeById(idPurpose);
+    }
+
 
 }
